requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
pytest
pytest-cov
pytest-asyncio
python-multipart
httpx
flake8
pylint